% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_solutions_wrapper.R
\name{autoload_environments}
\alias{autoload_environments}
\title{Load solutions to parent environment automatically}
\usage{
autoload_environments(
  instructor_file,
  student_file,
  student_file_MarkUs,
  markus_environment_variable = "MARKUS_GROUP",
  print_variables = FALSE
)
}
\arguments{
\item{instructor_file}{The solution file name (and path, relative to the project's root folder for the instructor solutions. Valid filetypes are .Rmd and .R}

\item{student_file}{The student submission file. Valid filetypes are .Rmd and .R}

\item{student_file_MarkUs}{The expected file name for student submission on MarkUs (and path, relative to the project's root folder. Valid filetypes are .Rmd and .R}

\item{markus_environment_variable}{The name of the environmental variable on MarkUs}

\item{print_variables}{Boolean value indicating whether the names of the variables loaded from the specified filepath should be printed. Defaults to false}
}
\value{
Print successful messages.
}
\description{
Load solutions to parent environment automatically
}
