% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotests.R
\name{correctArgsTest}
\alias{correctArgsTest}
\title{Test if the arguments have correct data types}
\usage{
correctArgsTest(variableName, student_environment, instructor_environment)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{student_environment}{A list of all variables in the environment from the student's submission}

\item{instructor_environment}{A list of all variables in the environment from the solution file}
}
\value{
an error message will be provided if none of the tests passed
}
\description{
Test if the arguments have correct data types
}
