% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotests.R
\name{correctSolnTest}
\alias{correctSolnTest}
\title{Test Correct Solution}
\usage{
correctSolnTest(studentSoln, actualSoln, order = TRUE, type)
}
\arguments{
\item{studentSoln}{The student's solution loaded from their assignment}

\item{actualSoln}{The actual solution for the question}

\item{order}{Whether the order of the data matters for lists. Default is TRUE.}

\item{type}{The type of data. Options are: scalar, list, dataframe}
}
\value{
Message for a successful test or an error message if fails
}
\description{
Tests to see if the student solution is equal to the actual solution
}
