% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotests.R
\name{dataTypeTest}
\alias{dataTypeTest}
\title{Test Data Type}
\usage{
dataTypeTest(
  variableName,
  variables,
  studentSoln,
  datatype,
  error_message = NULL
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{variables}{List of variables in the student solution environment}

\item{studentSoln}{The student's solution loaded from their assignment}

\item{datatype}{A string representing the datatype expected for the solution}

\item{error_message}{A function that will generate the appropriate error message as a string. Default is NULL and will use preset error message.}
}
\value{
Message for a successful test or an error message if fails
}
\description{
Tests to see if the student's solution is of the correct data type
}
