% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{extractData}
\alias{extractData}
\title{Data Extraction}
\usage{
extractData(file, sep = ",", ...)
}
\arguments{
\item{file}{Filename of the full data set}

\item{sep}{The field separator character. Values on each line of the \
file are separated by this character. If sep="", the separator is 'white space',
that is one or more spaces, tabs, newlines, or carriage returns.
Permitted values are: ",", "", and tab (backslash t); defaults to ",".}

\item{...}{Further arguments to be passed to the function}
}
\value{
The data extracted from the file as a data.frame
}
\description{
Extract the data from the file
}
