% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate Data}
\usage{
generateData(distribution, n, seedNum, ...)
}
\arguments{
\item{distribution}{Distribution to generate data from.
Options are: 'beta', 'binom', 'chisq', 'exp', 'f', 'gamma', 'geom', 'hyper',
'logis', 'lnorm', 'nbinom', 'norm', 'pois', 't', 'unif', 'weibull'
Example: [0,1] for mean and standard deviation of Normal distribution}

\item{n}{Size of dataset}

\item{seedNum}{Seed to randomly select data from data set}

\item{...}{The remaining parameters to set based on random distribution}
}
\value{
Data frame with randomly generated data from specified probability distribution
}
\description{
Generate data from a random probability distribution
}
\examples{
vals1 <- generateData('norm', 10, 25, mean=0, sd=1)
vals2 <- generateData('binom', 10, 25, size=2, prob=0.5)
}
