% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_running_on_MarkUs.R
\name{is_running_on_MarkUs}
\alias{is_running_on_MarkUs}
\title{Test the existence of the environmental variable MARKUS_GROUP if the file is on MarkUs}
\usage{
is_running_on_MarkUs(markus_environment_variable = "MARKUS_GROUP")
}
\arguments{
\item{markus_environment_variable}{The name of the environmental variable on MarkUs}
}
\value{
A string of the environmental variable, empty if the environmental
does not exist.
}
\description{
Test the existence of the environmental variable MARKUS_GROUP if the file is on MarkUs
}
