% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_solutions.R
\name{load_solutions}
\alias{load_solutions}
\title{Loads solutions file and return a list of variables in the environment}
\usage{
load_solutions(file, print_variables = FALSE)
}
\arguments{
\item{file}{The file name (and path, relative to the project's root folder. Valid filetypes are .Rmd and .R}

\item{print_variables}{Boolean value indicating whether the names of the variables loaded from the specified filepath should be printed. Defaults to false}
}
\value{
List of all variables in the instructor solutions
}
\description{
Loads solutions file and return a list of variables in the environment
}
