% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_solutions.R
\name{load_student_submission}
\alias{load_student_submission}
\title{Loads student submission file (locally or from MarkUs) and return a list of variables in the environment}
\usage{
load_student_submission(
  file,
  file_MarkUs = NULL,
  markus_environment_variable = "MARKUS_GROUP",
  print_variables = FALSE
)
}
\arguments{
\item{file}{The local file name (and path, relative to the project's root folder for the instructor solutions. Valid filetypes are .Rmd and .R}

\item{file_MarkUs}{The expected file name for student submissions on MarkUs (and path, relative to the project's root folder. Valid filetypes are .Rmd and .R}

\item{markus_environment_variable}{The name of the environmental variable on MarkUs}

\item{print_variables}{Boolean value indicating whether the names of the variables loaded from the specified filepath should be printed. Defaults to false}
}
\value{
List of all variables in the student submission
}
\description{
Loads student submission file (locally or from MarkUs) and return a list of variables in the environment
}
