% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMarkUs-package.R
\docType{package}
\name{RMarkUs-package}
\alias{RMarkUs-package}
\title{RMarkUs: Autograding Package for MarkUs}
\description{
This package will allow for autograding capabilities for quantitative assessments using MarkUs.
}
\author{
\strong{Maintainer}: Siqi Zheng \email{timothy.zheng@mail.utoronto.ca}

Authors:
\itemize{
  \item Nathalie Moon [funder]
  \item Kanika Chopra
}

}
\keyword{internal}
