% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{run_function}
\alias{run_function}
\title{Probability Distribution}
\usage{
run_function(fct, ...)
}
\arguments{
\item{fct}{Probability distribution function in R as a string}

\item{...}{Parameters for function}
}
\value{
List of random values generated from probability distribution
}
\description{
Helper function to run the distribution function using the string name
}
\examples{
vals1 <- run_function('rnorm', 10, 0, 1)
vals2 <- run_function('rbinom', 10, 2, 0.5)
}
