% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{sourceList}
\alias{sourceList}
\title{Source List}
\usage{
sourceList(path)
}
\arguments{
\item{path}{Filepath}
}
\value{
A list of variables from specified environment
}
\description{
Extracts all variables from the R file into a list
}
