% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{subsetData}
\alias{subsetData}
\title{Random Subset}
\usage{
subsetData(file, sep, size, variables = "all", seedNum, ...)
}
\arguments{
\item{file}{Filename of full dataset}

\item{sep}{The field separator character. Values on each line of the \
file are separated by this character. If sep="", the separator is 'white space',
that is one or more spaces, tabs, newlines, or carriage returns.
Permitted values are: ",", "", and tab (backslash t); defaults to ",".}

\item{size}{Number of rows to choose for the subset.}

\item{variables}{Variables to select from data set.
DEFAULT is all.}

\item{seedNum}{Seed to randomly select data from data set}

\item{...}{Further arguments to be passed to the function}
}
\value{
The random subset of data from the full dataset
}
\description{
Randomly subset data from a full dataset
}
