% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pdf_exists.R
\name{test_pdf_exists}
\alias{test_pdf_exists}
\title{Check for pdf}
\usage{
test_pdf_exists(filename = NULL, error_message = NULL)
}
\arguments{
\item{filename}{Optional parameter for the name of the pdf file you're interested in checking for.
If none is provided, the function will just check that a pdf file exists}

\item{error_message}{A function that will generate the appropriate error message as a string for if the data type is not correct.
Default is NULL and will use preset error message.}
}
\value{
A list of variables from specified environment
}
\description{
Test to check if pdf file was submitted
}
