% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_student_number.R
\name{test_student_number}
\alias{test_student_number}
\title{Wrapper for testing student number, essentially testScalar on student number}
\usage{
test_student_number(
  varname = "STUDENT_NUMBER",
  student_environment = student_environment,
  instructor_environment = instructor_environment
)
}
\arguments{
\item{varname}{default is the student number}

\item{student_environment}{A list of all variables in the environment from the student's submission}

\item{instructor_environment}{A list of all variables in the environment from the solution file}
}
\value{
It will print that every test has passed.
}
\description{
Wrapper for testing student number, essentially testScalar on student number
}
