% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testDataFrame}
\alias{testDataFrame}
\title{DataFrame Tests}
\usage{
testDataFrame(
  variableName,
  student_environment = student_environment,
  instructor_environment = instructor_environment,
  check_correct = TRUE,
  correct_error_msg = NULL,
  check_present = TRUE,
  present_error_msg = NULL,
  check_size = TRUE,
  size_error_msg = NULL,
  check_attributes = FALSE,
  attributes_error_msg = NULL,
  check_class = TRUE,
  class_error_msg = NULL,
  col_order = TRUE
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{student_environment}{A list of all variables in the environment from the student's submission}

\item{instructor_environment}{A list of all variables in the environment from the solution file}

\item{check_correct}{Boolean indicating whether a test will check if the solution is correct.
Default is TRUE.}

\item{correct_error_msg}{A function that will generate the appropriate error message as a string if the value is incorrect.
Default is NULL and will use preset error message.}

\item{check_present}{Boolean indicating whether a test will check if the variable exists.
Default is TRUE.}

\item{present_error_msg}{A function that will generate the appropriate error message as a string for if variable does not exist.
Default is NULL and will use preset error message.}

\item{check_size}{Boolean indicating whether a test will check if the size of the dataframe is correct.
Default is TRUE.}

\item{size_error_msg}{A function that will generate the appropriate error message as a string if the size is incorrect.
Default is NULL and will use preset error message.}

\item{check_attributes}{Boolean indicating whether a test will check if the attributes of the dataframe are correct.
Default is TRUE.}

\item{attributes_error_msg}{A function that will generate the appropriate error message as a string if the attributes of the dataframe are incorrect.
Default is NULL and will use preset error message.}

\item{check_class}{Boolean indicating whether a test will check if the variable classes of the dataframe are correct.
Default is TRUE.}

\item{class_error_msg}{A function that will generate the appropriate error message as a string if the classes of the dataframe are incorrect.
Default is NULL and will use preset error message.}

\item{col_order}{Boolean indicating if order matters in correctness.
Default is FALSE}
}
\value{
Error message if one exists, otherwise will print that every test has passed.
}
\description{
Completes all of the tests for a DataFrame, where the user can pass in the full environments from student and instructor solutions
}
