% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testGroupNumber}
\alias{testGroupNumber}
\title{Compare if a particular column has the correct group and whether the number of
observation in each group is correct.}
\usage{
testGroupNumber(
  variableName,
  student_environment = student_environment,
  instructor_environment = instructor_environment,
  subset = FALSE,
  column_name = NULL
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{student_environment}{A list of all variables in the environment from the student's submission}

\item{instructor_environment}{A list of all variables in the environment from the solution file}

\item{subset}{Boolean indicating whether another column needs to be selected.
Use if an entire dataframe is passed to variableName.
Default is False}

\item{column_name}{A string of the column name}
}
\value{
Error message if one exists, otherwise will print that every test has passed.
}
\description{
Compare if a particular column has the correct group and whether the number of
observation in each group is correct.
}
