% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testLinearModel}
\alias{testLinearModel}
\title{Compare if 2 linear models are the same}
\usage{
testLinearModel(
  variableName,
  student_environment,
  instructor_environment,
  check_correct = TRUE,
  correct_error_msg = NULL,
  check_present = TRUE,
  present_error_msg = NULL,
  AIC_compare = FALSE,
  round_precision = 3
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{student_environment}{A list of all variables in the environment from the student's submission}

\item{instructor_environment}{A list of all variables in the environment from the solution file}

\item{check_correct}{Boolean indicating whether a test will check if the solution is correct.
Default is TRUE.}

\item{correct_error_msg}{A function that will generate the appropriate error message as a string if the value is incorrect.
Default is NULL and will use preset error message.}

\item{check_present}{Boolean indicating whether a test will check if the variable exists.
Default is TRUE.}

\item{present_error_msg}{A function that will generate the appropriate error message as a string for if variable does not exist.
Default is NULL and will use preset error message.}

\item{AIC_compare}{Optional parameter that enables AIC comparison of models. Default is FALSE.}

\item{round_precision}{The number of decimal places RSS and df will be rounded. Default is 3.}
}
\value{
Error message if one exists, otherwise will print that every test has passed.
}
\description{
Compare if 2 linear models are the same
}
