% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testScalar}
\alias{testScalar}
\title{Scalar Tests}
\usage{
testScalar(variableName, variables, studentSoln, actualSoln, datatype)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{variables}{List of variables in environment}

\item{studentSoln}{The student's solution loaded from their assignment}

\item{actualSoln}{The actual solution for the question}

\item{datatype}{A string representing the datatype expected for the solution}
}
\value{
Error message if one exists, otherwise will print that every test has passed.
}
\description{
Completes all of the basic tests for a scalar value
}
