% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testScalar_raw}
\alias{testScalar_raw}
\title{testScalar_raw}
\usage{
testScalar_raw(
  variableName,
  variables,
  studentSoln,
  actualSoln,
  check_present = TRUE,
  present_error_msg = NULL,
  check_datatype = FALSE,
  data_error_msg = NULL,
  check_correct = TRUE,
  correct_error_msg = NULL,
  datatype = NULL
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{variables}{Vector listing the names of variables in the student's environment}

\item{studentSoln}{The student's answer, loaded from their submission}

\item{actualSoln}{The actual answer, loaded from the solution file}

\item{check_present}{Boolean indicating whether a test will check if the variable exists.
Default is TRUE.}

\item{present_error_msg}{A function that will generate the appropriate error message as a string for if variable does not exist.
Default is NULL and will use preset error message.}

\item{check_datatype}{Boolean indicating whether a test will check if the data type is correct.
Default is FALSE.}

\item{data_error_msg}{A function that will generate the appropriate error message as a string for if the data type is not correct.
Default is NULL and will use preset error message.}

\item{check_correct}{Boolean indicating whether a test will check if the solution is correct.
Default is TRUE.}

\item{correct_error_msg}{A function that will generate the appropriate error message as a string if the value is incorrect.
Default is NULL and will use preset error message.}

\item{datatype}{Optional argument (string) for data type expected for solution if check_datatype is TRUE.
Accepts values "numeric", "character", "logical", and "double" (if numeric, then tests pass if solution is either double or integer)}
}
\value{
Error message if one exists, otherwise will print that every test has passed.
}
\description{
Completes all of the basic tests for a single (scalar) value (that is, a vector of length 1), where the user can directly input values to be compared
}
