% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testVector}
\alias{testVector}
\title{Vector Tests}
\usage{
testVector(
  variableName,
  variables,
  studentSoln,
  actualSoln,
  check_present = TRUE,
  present_error_msg = NULL,
  check_correct = TRUE,
  correct_error_msg = NULL,
  check_length = TRUE,
  size_error_msg = NULL,
  check_datatype = FALSE,
  datatype = NULL,
  type = "vector",
  order = TRUE,
  data_error_msg = NULL
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{variables}{List of variables in environment}

\item{studentSoln}{The student's solution loaded from their assignment}

\item{actualSoln}{The actual solution for the question}

\item{check_present}{Boolean indicating whether a test will check if the variable exists.
Default is TRUE.}

\item{present_error_msg}{A function that will generate the appropriate error message as a string for if variable does not exist.
Default is NULL and will use preset error message.}

\item{check_correct}{Boolean indicating whether a test will check if the solution is correct.
Default is TRUE.}

\item{correct_error_msg}{A function that will generate the appropriate error message as a string if the value is incorrect.
Default is NULL and will use preset error message.}

\item{check_length}{Boolean indicating whether a test will check if the length of the vector is correct.
Default is TRUE.}

\item{size_error_msg}{A function that will generate the appropriate error message as a string if the size is incorrect.
Default is NULL and will use preset error message.}

\item{check_datatype}{Boolean indicating whether a test will check if the data type is correct.
Default is FALSE.}

\item{datatype}{Optional argument (string) for data type expected for solution if check_datatype is TRUE.
Options are `numeric`, `character`, `logical`}

\item{type}{Optional argument (string) for data type expected for solution if check_datatype is TRUE.
DEFAULT is vector. Options are `vector`, `scalar`, `list`}

\item{order}{Boolean indicating if order matters in correctness.
Default is TRUE.}

\item{data_error_msg}{A function that will generate the appropriate error message as a string for if the data type is not correct.
Default is NULL and will use preset error message.}
}
\value{
Error message if one exists, otherwise will print that every test has passed.
}
\description{
Completes all of the tests for a vector
}
