% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotests.R
\name{correctSolnTest}
\alias{correctSolnTest}
\title{Test Correct Solution}
\usage{
correctSolnTest(
  variableName,
  variables,
  studentSoln,
  actualSoln,
  order = TRUE,
  type,
  error_message = NULL
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{studentSoln}{The student's solution loaded from their assignment}

\item{actualSoln}{The actual solution for the question}

\item{order}{Whether the order of the data matters for lists. Default is TRUE.}

\item{type}{The type of data. Options are: scalar, vector, dataframe}

\item{error_message}{A function that will generate the appropriate error message as a string. Default is NULL and will use preset error message.}
}
\value{
Message for a successful test or an error message if fails
}
\description{
Tests to see if the student solution is equal to the actual solution
}
