% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotests.R
\name{dataTypeTest}
\alias{dataTypeTest}
\title{Test Data Type}
\usage{
dataTypeTest(studentSoln, datatype)
}
\arguments{
\item{studentSoln}{The student's solution loaded from their assignment}

\item{datatype}{A string representing the datatype expected for the solution}
}
\value{
Message for a successful test or an error message if fails
}
\description{
Tests to see if the student's solution is of the correct data type
}
