% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{extractData}
\alias{extractData}
\title{Data Extraction}
\usage{
extractData(filename, filetype)
}
\arguments{
\item{filename}{Filename of full data set}

\item{filetype}{Type of file, i.e. csv, tab-delim, space-delim}
}
\value{
The data extracted from the file as a data.frame
}
\description{
Extract the data from the file
}
