% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMarkUs.R
\docType{package}
\name{RMarkUs}
\alias{RMarkUs}
\title{Autograding Package for MarkUs}
\description{
This package will allow for autograding capabilities for quantitative assessments using MarkUs.
}
\author{
Kanika Chopra \email{kanikadatt@gmail.com}
}
