% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{subsetData}
\alias{subsetData}
\title{Random Subset}
\usage{
subsetData(filename, filetype, n, variables = "all", seedNum)
}
\arguments{
\item{filename}{Filename of full dataset}

\item{filetype}{Type of file, i.e. csv files, text files, table}

\item{n}{Size of dataset}

\item{variables}{Variables to select from data set.
DEFAULT is all.}

\item{seedNum}{Seed to randomly select data from data set}
}
\value{
The random subset of data from the full dataset
}
\description{
Randomly subset data from a full dataset
}
