% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testScalar}
\alias{testScalar}
\title{testScalar}
\usage{
testScalar(
  variableName,
  student_environment,
  instructor_environment,
  check_present = TRUE,
  present_error_msg = NULL,
  check_datatype = TRUE,
  data_error_msg = NULL,
  check_correct = TRUE,
  correct_error_msg = NULL,
  datatype = NULL,
  simplify_datatype = FALSE
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{student_environment}{A list of all variables in the environment from the student's submission}

\item{instructor_environment}{A list of all variables in the environment from the solution file}

\item{check_present}{Boolean indicating whether a test will check if the variable exists.
Default is TRUE.}

\item{present_error_msg}{A function that will generate the appropriate error message as a string for if variable does not exist.
Default is NULL and will use preset error message.}

\item{check_datatype}{Boolean indicating whether a test will check if the data type is correct.
Default is FALSE.}

\item{data_error_msg}{A function that will generate the appropriate error message as a string for if the data type is not correct.
Default is NULL and will use preset error message.}

\item{check_correct}{Boolean indicating whether a test will check if the solution is correct.
Default is TRUE.}

\item{correct_error_msg}{A function that will generate the appropriate error message as a string if the value is incorrect.
Default is NULL and will use preset error message.}

\item{datatype}{Optional argument (string) for data type expected for solution if check_datatype is TRUE.
If NULL, will check against the data type of the solution
Accepts values "numeric", "character", "logical", and "double" (if numeric, then tests pass if solution is either double or integer)}

\item{simplify_datatype}{Boolean indicating whether datatype tests will pass if a student passes a tibble when the solutions expect a primitive instead. Defaults to FALSE}
}
\value{
Error message if one exists, otherwise will print that every test has passed.
}
\description{
Completes all of the basic tests for a single (scalar) value (that is, a vector of length 1), where the user can pass in the full environments from student and instructor solutions
}
