% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testScalar}
\alias{testScalar}
\title{Scalar Tests}
\usage{
testScalar(
  variableName,
  variables,
  studentSoln,
  actualSoln,
  check_datatype = FALSE,
  datatype = "double"
)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{variables}{List of variables in environment}

\item{studentSoln}{The student's solution loaded from their assignment}

\item{actualSoln}{The actual solution for the question}

\item{check_datatype}{Boolean indicating whether data type will be checked.
Default is FALSE.}

\item{datatype}{Optional argument (string) for data type expected for solution if check_datatype is TRUE.
DEFAULT is double.}
}
\value{
Error message if one exists, otherwise will print that every test has passed.
}
\description{
Completes all of the basic tests for a scalar value
}
