% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotests.R
\name{variableClassTest}
\alias{variableClassTest}
\title{Test Variable Class}
\usage{
variableClassTest(studentSoln, actualSoln, datatype, error_message = NULL)
}
\arguments{
\item{studentSoln}{The student's solution loaded from their assignment}

\item{actualSoln}{The actual solution for the question}

\item{datatype}{A string representing the datatype expected for the solution}

\item{error_message}{A function that will generate the appropriate error message as a string. Default is NULL and will use preset error message.}
}
\value{
Message for a successful test or an error message if fails
}
\description{
Tests to see if the classes of the DataFrame columns are the same
}
