% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotests.R
\name{variableExistsTest}
\alias{variableExistsTest}
\title{Test if Variable Exists}
\usage{
variableExistsTest(variableName, variables, error_message = NULL)
}
\arguments{
\item{variableName}{The name of the variable in question}

\item{variables}{List of variables in environment}

\item{error_message}{A function that will generate the appropriate error message as a string. Default is NULL and will use preset error message.}
}
\value{
Message for a successful test or an error message if fails
}
\description{
Tests to see if the variable for a question exists in the R environment
}
